+++ database/migrations/2026_09_14_234034_create_commande.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->integer('nb_actif')->nullable();
             $table->date('date_commande')->nullable();
 
-            $table->unsignedInteger('id_etat');
-            $table->unsignedInteger('id_emplacement_prevu')->nullable();
+            $table->unsignedBigInteger('id_etat');
+            $table->unsignedBigInteger('id_emplacement_prevu')->nullable();
 
             $table->foreign('id_etat')->references('id')->on('etat');
             $table->foreign('id_emplacement_prevu')->references('id')->on('emplacement');
