+++ database/migrations/2029_09_14_234101_create_client.php
@@ -20,10 +20,10 @@ return new class extends Migration
             $table->boolean('emplacement_manuel')->default(FALSE);
             $table->boolean('inactif')->default(FALSE);
 
-            $table->unsignedInteger('id_poste')->nullable();
-            $table->unsignedInteger('id_type_client');
-            $table->unsignedInteger('id_actif')->nullable();
-            $table->unsignedInteger('id_emplacement')->nullable();
+            $table->unsignedBigInteger('id_poste')->nullable();
+            $table->unsignedBigInteger('id_type_client');
+            $table->unsignedBigInteger('id_actif')->nullable();
+            $table->unsignedBigInteger('id_emplacement')->nullable();
 
             $table->foreign('id_poste')->references('id')->on('poste');
             $table->foreign('id_type_client')->references('id')->on('type_client');
